Make astar and BinaryHeap use double costs instead of float (matches related code)
There's also a new CostlyConnection that stores a cost, and DefaultGraph can take a double[][] that was produced for the old AStarSearch to assign different costs to different nodes.

+++ squidlib-util/src/main/java/squidpony/squidai/astar/Connection.java
@@ -28,7 +28,7 @@ package squidpony.squidai.astar;
 public interface Connection<N> {
 
 	/** Returns the non-negative cost of this connection */
-	float getCost();
+	double getCost();
 
 	/** Returns the node that this connection came from */
 	N getFromNode();
